Clean up yarn idempotency mess
Yarn quires a yarn.lock file to be committed to git, but when installing dependencies (yarn install), it still modified the yarn.lock file.  That causes the git repo to become dirty, so git refuses to cleanly pull new revisions (or cleanly check that no upstream changes are required).  So make sue the yarn.lock file is reset after yarn does its thing.